Added directives, sleepers, atomic integers.  Killscore modifiers are now proportional to need (to kill or create a thread).
The threadpool doesn't *actually* need to keep track of its threads -- they know how to find the job pool, and we can
keep track of the number of threads by tracking them with an AtomicInteger counter.  And instead of tracking and waking threads
manually using a mutex in the batch, we can use a "sleeper" (funny name I made up...) to put the current thread to sleep until
the batch is done.  Similar in interface and functionality to a ConditionVariable, but no mutexes involved.

Still plan on refactoring a bit more -- want at the very least one class per file, and to fix up the specs and get them running again.